Methods

+++ WebApplication1/Domains/Product.cs
@@ -17,7 +17,7 @@ namespace WebApplication1.Domains
         public virtual string Description { get; set; }
 
         public virtual decimal Price { get; set; }
-        public virtual decimal BasePrice { get; set; }
+        public virtual int BasePrice { get; set; }
         public virtual string Image { get; set; }
 
         public virtual int CategoryId { get; set; }
